Add a skeleton for server and peers
This  commit  includes a first simple
example that shows  how to send messa
ges over socket between server and cl
ient

Protocol's structure has beed defined
Each message will be  composed by two
parts: header + payload

HEADER: of two bytes. The first one is
the length of the payload , the second
one is the type  of message ( register
login, logout, search  and  so on ...)

<---1-byte----><-----1-byte----->

+-------------+-----------------+
| payload_len | type_of_message |
+-------------+-----------------+

PAYLOAD: it' s different  depending on
the type of messages exchanged. Please
check documentation for details.